Memory and collections update

+++ src/utilities/collections/equaility_comparer_ref.rs
@@ -1,4 +1,7 @@
+/// Trait for types that can compare and hash items by reference.
 pub trait RefEqualityComparer<T> {
-    fn hash(&self, item: &T) -> u64;
+    /// Computes a hash code for the given item.
+    fn hash(&self, item: &T) -> i32;
+    /// Checks if two items are equal.
     fn equals(&self, a: &T, b: &T) -> bool;
 }
